# patch/make git repo root finding faster