This PR is based on the changes in #480, and improves upon the git repo root finding.

Previously we have relied on `git` cmd to be installed to find the repo root, which required spawning a new process, which turned out to be rather slow. This PR replaces that implementation with traversing the file tree.

I tested this both on macOS and Windows, the behavior doesn't seem to change in any way.

## Micro-benchmark
I have ran a Go benchmark for this method specifically, the change is **very** significant:
```
                    │    old.txt     │               new.txt               │
                    │     sec/op     │   sec/op     vs base                │
FindRepoFromPath-12   5032.729µ ± 8%   4.520µ ± 9%  -99.91% (p=0.000 n=10)

                    │    old.txt    │               new.txt                │
                    │     B/op      │     B/op      vs base                │
FindRepoFromPath-12   54.342Ki ± 0%   1.627Ki ± 0%  -97.01% (p=0.000 n=10)

                    │   old.txt   │              new.txt               │
                    │  allocs/op  │ allocs/op   vs base                │
FindRepoFromPath-12   130.00 ± 0%   16.00 ± 0%  -87.69% (p=0.000 n=10)
```

99% faster than the previous version with 97% less memory.

## Impact on asset parsing
On top of the improvements in #480, this seems to deliver an additional ~30% on my MBP.

<img width="316" alt="image" src="https://github.com/user-attachments/assets/f92a0b83-2dc6-467b-84e1-4e0b4d074646" />

I tested this on a windows VM on my mac, the difference is 77%:

<img width="275" alt="image" src="https://github.com/user-attachments/assets/d06f184a-92dc-4f07-9348-80d12d941bc2" />